Fix build error in solven.f
The error:
solven.f:458:72:

  458 |          call rate(t9,rho,dtstar)
      |                                                                        1
Error: Missing actual argument for argument '_formal_3' at (1)

The (apparent) cause:
'rate' was called earlier in the file with two extra arguments. I'm
guessing the compiler silently changed the function signature??

N.b.: solven-ma.f was afflicted by the same problem, and is fixed here,
too.

+++ coolandburn.c
@@ -692,7 +692,7 @@ int prep_cool_burn(SPHbody *p, float tlo, float tup, int Gridpts, int Nel, int r
 float burning(SPHbody *p, float dt, int rank) {
 
     int i,j,l;
-    int partid;
+    long partid;
     double temp, rho, dt_cgs, ndens;
     double deltah;
     double *molfrac;
